Makefile: Cleanup LD_FLAGS
The various options set in LD_FLAGS don't have any effect in the linker,
they all control the compilation stage.

So set LD_FLAGS to empty.

This does require a slight change to the test_build recipe as this does
some compilation and linking. Use CC (although it's the same as LD) and
use CC_FLAGS (as well as LD_FLAGS, although it's currently empty).