client/c: Add IPv6 support
This adds IPv6 support to the C client. It makes use of getaddrinfo(3).
This also allows to use hostnames instead of IP addresses.

We drop the call to bind(2) and just let the kernel handle it.

We store the port and server IP address as const char *'s in the config
structure so these can just be passed directly to getaddrinfo() and the
tests were updated to reflect that.